:bug: fix: boardStar Type change double -> int

+++ taste_tour/lib/src/model/feed_model.dart
@@ -7,7 +7,7 @@ class FeedModel {
 
   int? boardNumber;
   String? boardTitle;
-  double? boardStar;
+  int? boardStar;
   String? boardCategory;
   String? boardStoreLocation;
   String? boardContent;
